Limpieza: corregir los ultimos warnings W4 del build de tests
- C4100 en SaveSystem.cpp: el parametro useDeltaCompression esta reservado en
  la firma pero sin implementar; se anonimiza.
- C4267 (size_t a uint8_t) en el bit-packing de PalettedStorage.h: la resta
  64 - bitPosition siempre cabe en 8 bits (bitPosition < 64); cast explicito.

Con esto el build de tests con nivel W4 queda sin warnings, y el del juego
solo conserva el C4005 de APIENTRY por el orden GLFW/windows.h, preexistente.

Nota: el borrado de src/TerrainGeneration (10 archivos muertos, ~143 KB, cero
referencias) entro por descuido en el commit anterior (2c988c1) junto con la
iluminacion: estaba en el indice desde antes. El contenido del arbol es el
correcto; solo quedo repartido en dos commits.

Verificado: compilan juego y tests, ctest 1/1 (2830 aserciones).

+++ src/PalettedStorage.h
@@ -71,8 +71,9 @@ private:
             return (uint16_t)((indices[arrayIndex] >> bitPosition) & mask);
         } else {
             // Caso complejo: el índice está dividido entre dos uint64_t
-            uint8_t bitsInFirst = 64 - bitPosition;
-            uint8_t bitsInSecond = bitsPerIndex - bitsInFirst;
+            // (bitPosition < 64, así que la resta cabe de sobra en 8 bits)
+            uint8_t bitsInFirst = (uint8_t)(64 - bitPosition);
+            uint8_t bitsInSecond = (uint8_t)(bitsPerIndex - bitsInFirst);
 
             uint64_t lowBits = (indices[arrayIndex] >> bitPosition) & ((1ULL << bitsInFirst) - 1);
             uint64_t highBits = (indices[arrayIndex + 1] & ((1ULL << bitsInSecond) - 1));
@@ -100,8 +101,9 @@ private:
             indices[arrayIndex] = (indices[arrayIndex] & ~(mask << bitPosition)) | (value << bitPosition);
         } else {
             // Caso complejo: dividido entre dos uint64_t
-            uint8_t bitsInFirst = 64 - bitPosition;
-            uint8_t bitsInSecond = bitsPerIndex - bitsInFirst;
+            // (bitPosition < 64, así que la resta cabe de sobra en 8 bits)
+            uint8_t bitsInFirst = (uint8_t)(64 - bitPosition);
+            uint8_t bitsInSecond = (uint8_t)(bitsPerIndex - bitsInFirst);
 
             uint64_t maskFirst = ((1ULL << bitsInFirst) - 1) << bitPosition;
             uint64_t maskSecond = (1ULL << bitsInSecond) - 1;
